refactor: movimentação dos inputs para uma pasta dedicada
