+++ frontend/src/screens/auth/LoginScreen.tsx
@@ -3,8 +3,8 @@ import { View, ScrollView } from "react-native";
 import { LogoSVG } from "../../components/Logo";
 import styles from "./styles";
 import { NavigationProp } from "@react-navigation/native";
-import { Input } from "../../components/Input";
-import { PasswordInput } from "../../components/PasswordInput";
+import { Input } from "../../components/inputs/Input";
+import { PasswordInput } from "../../components/inputs/PasswordInput";
 import { ScreenWrapper } from "../../components/ScreenWrapper";
 import { SubmitButton } from "../../components/buttons/SubmitButton";
 import { request } from "../../libs/axios/Request";
