+++ frontend/src/screens/auth/SignupScreen.tsx
@@ -4,8 +4,8 @@ import { LogoSVG } from "../../components/Logo";
 import styles from "./styles";
 import { NavigationProp } from "@react-navigation/native";
 import { ScreenWrapper } from "../../components/ScreenWrapper";
-import { Input } from "../../components/Input";
-import { PasswordInput } from "../../components/PasswordInput";
+import { Input } from "../../components/inputs/Input";
+import { PasswordInput } from "../../components/inputs/PasswordInput";
 import { SubmitButton } from "../../components/buttons/SubmitButton";
 import { request } from "../../libs/axios/Request";
 import { ScreenTitle } from "../../components/ScreenTitle";
